Experimental: async await on other file descriptors
This sets up epoll to wait on an arbitrary file descriptor using
EPOLLONESHOT; useful, for instance, if implementing asynchronous
database access using the MariaDB client libraries in non-blocking
mode.

This has not been tested; it's just an API experiment.  It's likely
that the connection file descriptor will need to be changed to some
kind of "suspended" state (similar to when it's suspended by a
timer).